[#655][refactor] extract preview sections to components
