+++ packages/lib/shared/utils/chart.helper.ts
@@ -1,0 +1,18 @@
+export function buildMarkline(id: string, start: Date, end: Date, height: number) {
+  return {
+    id,
+    type: 'line',
+    data: [
+      [start, height],
+      [end, height],
+    ],
+    lineStyle: {
+      color: 'grey',
+      type: 'dashed',
+      width: 1,
+      cap: 'round' as const,
+      join: 'round' as const,
+    },
+    showSymbol: false,
+  }
+}
